Summary
Reformatted the UI for all of the forms associated with issues and facilities.

Renamed the nested "Facilities" option in the sidebar to "Campus Directory."

Added the ability to see the rooms and floors for each of the facilities recorded in the system. Also added the ability to add another floor, if the user is an admin.